feat: refactor permissions enum with proper bit allocation
- Fix bit position conflicts and organize permissions logically
- Add missing permissions and resolve duplicates
- Improve role hierarchy: BasicUser → TeamMember → TeamManager → OrgAdmin → OrgOwner → SuperAdmin → SystemAdmin
- Update controller permissions for granular access control
- Ensure scalable structure with unique bit positions 0-65

+++ Shortly.Infrastructure/DbContexts/Configurations/OrganizationMemberConfiguration.cs
@@ -15,7 +15,7 @@ public class OrganizationMemberConfiguration : IEntityTypeConfiguration<Organiza
         // Properties configuration
         builder.Property(e => e.RoleId)
             .HasConversion<byte>()
-            .HasDefaultValue((byte)enUserRole.Member);
+            .HasDefaultValue(enUserRole.Member);
 
         builder.Property(e => e.CustomPermissions)
             .HasConversion<long>().HasDefaultValue(enPermissions.None);
